+++ src/AppModule.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { BooksModule } from './book/book.module';
-import { typeORMConfig } from './configs/typeorm.config';
+import { BooksModule } from './book/BookModule';
+import { typeORMConfig } from './config/TypeOrmConfig';
 
 @Module({
   imports: [
