+++ src/book/dto/GetBooksQueryDto.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty } from "class-validator";
+
+
+export class GetBooksQueryDto {
+  @IsNotEmpty()
+  bookQuery: string;
+}
