+++ src/db/TypeOrmExModule.ts
@@ -1,7 +1,7 @@
 import { DynamicModule, Provider } from "@nestjs/common";
 import { getDataSourceToken } from "@nestjs/typeorm";
 import { DataSource } from "typeorm";
-import { TYPEORM_EX_CUSTOM_REPOSITORY } from "./typeorm-ex.decorator";
+import { TYPEORM_EX_CUSTOM_REPOSITORY } from "./TypeOrmExDecorator";
 
 export class TypeOrmExModule {
   public static forCustomRepository<T extends new (...args: any[]) => any>(repositories: T[]): DynamicModule {
