+++ src/db/time/BaseTimeEntity.ts
@@ -4,9 +4,9 @@ export abstract class BaseTimeEntity {
 	@PrimaryGeneratedColumn()
   	id: number;
 
-	@CreateDateColumn({ type: 'timestamptz' })
+	@CreateDateColumn({ type: 'timestamp' })
 	createdAt: Date;
 
-	@UpdateDateColumn({ type: 'timestamptz' })
+	@UpdateDateColumn({ type: 'timestamp' })
 	updatedAt: Date;
 }
