+++ src/main.ts
@@ -1,6 +1,6 @@
-import 'dotenv/config';
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import 'dotenv/config';
+import { AppModule } from './AppModule';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
